fix: удалить промисы из Card

+++ src/pages/index.js
@@ -157,16 +157,53 @@ Promise.all([api.getUserInfo(), api.getCardsInfo()])
             })
           }
         },
-        {//Добавить лайк.
-          addLike: () => {
-            return api.addLike(item._id);
+        {
+          handleLikeClick: (isLiked) => {
+            if (isLiked) {
+              api.deleteLike(item._id)
+                .then((response) => {
+                  newCard.toggleLike(response.likes)
+                })
+                .catch((error) => {
+                  console.log(`Ошибка: ${error}`);
+                });
+            } else {
+              api.addLike(item._id)
+                .then((response) => {
+                  newCard.toggleLike(response.likes)
+                })
+                .catch((error) => {
+                  console.log(`Ошибка: ${error}`);
+                })
+            }
           }
-        },
-        {//Удалить лайк.
-          deleteLike: () => {
-            return api.deleteLike(item._id);
-          }
-        });
+        }
+        // {//Добавить лайк.
+        //   addLike: () => {
+        //     return api.addLike(item._id)
+        //       .then((response) => {
+        //         this._counter.textContent = response.likes.length;
+        //         this._isLiked = !this._isLiked;
+        //       })
+        //       .then(() => {
+        //         this._like.classList.toggle('card__like_active');
+        //       })
+        // .catch((error) => {
+        //   console.log(`Ошибка: ${error}`);
+        // });
+        //   }
+        // },
+        // {//Удалить лайк.
+        //   deleteLike: () => {
+        //     return api.deleteLike(item._id)
+        //       .then((response) => {
+        //         return response.likes;
+        //       })
+        //       .catch((error) => {
+        //         console.log(`Ошибка: ${error}`);
+        //       });
+        //   }
+      );
       //2 наполнить карточку.
       const cardElement = newCard.generateCard(isIdEqual(item));
       //3 вставить карточку в начало списка.
